+++ src/core/paving/tubex_TubePaving.cpp
@@ -34,7 +34,7 @@ namespace tubex
     bool is_out = v_t_inv.empty();
     bool is_in = false;
 
-    for(int i = 0 ; i < v_t_inv.size() && !is_in ; i++)
+    for(size_t i = 0 ; i < v_t_inv.size() && !is_in ; i++)
     {
       const Slice **s = new const Slice*[size()];
       for(int j = 0 ; j < size() ; j++)
